Split up tests into separate modules grouped by concern
